rip98d: Perform a crapectomy on hex2intrev().
hex2intrev is a hilariously complicated function which combine parsing
a hexadecimal integer in network byte order and swapping the result to
host byte order.

Move the endianess swapping to the caller which allows to remove double
endianess swapping.  Now that hex2intrev no longer deals with byte
order rename it to hex2int and replace its implementation with a simple
invocation of strtoul.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>