rxecho: Increase space allocated for network interface name to IFNAMSIZ.
The previous hardcoded value was ok when used with the old SOCK_PACKET
which was truncating network interface names to 14 bytes anyway.  When
support for the new-fangled PF_PACKET was added this became a bug.

While technically a buffer overflow it is something that could only be
exploited by root himself.  Even more, due to alignment rules used by all
Linux architectures the two bytes following the overflowed arrays would
be unused anyway because an int is following, so this would actually end
up working as intended, not as coded.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
Fixes: 5415c1237782